Added a name field and a back button to the game over menu
The student satisfaction score is now added to the leaderboard when the game ends
The leaderboard now uses doubles as the score rather than ints

+++ src/main/java/y111studios/screens/LeaderboardScreen.java
@@ -100,7 +100,7 @@ public class LeaderboardScreen extends ScreenWithBackground {
         leaderboardRows[i + 1] = new Label("-- empty --", SKIN);
       } else {
         leaderboardRows[i] = new Label(((i / 2) + 1) + ". " + score.getName(), SKIN);
-        leaderboardRows[i + 1] = new Label(Integer.toString(score.getScore()), SKIN);
+        leaderboardRows[i + 1] = new Label(Double.toString(score.getScore()), SKIN);
       }
       table.row();
       table.add(leaderboardRows[i]).padTop(height * 0.05f);
