+++ src/main/java/y111studios/utils/Score.java
@@ -4,9 +4,9 @@ import lombok.Getter;
 
 public class Score {
   private @Getter String name;
-  private @Getter int score;
+  private @Getter double score;
 
-  public Score(String name, int score) {
+  public Score(String name, double score) {
     this.name = name;
     this.score = score;
   }
